modificada la estructura del proyecto. Agregados los paquetes dao, daoImpl, negocio y negocioImpl.

+++ WebContent/ListadoClientes.jsp
@@ -13,6 +13,9 @@
 	href="https://cdn.jsdelivr.net/npm/bootstrap@4.5.3/dist/css/bootstrap.min.css"
 	integrity="sha384-TX8t27EcRE3e/ihU7zmQxVncDAy5uIKz4rEkgIXeMed4M0jlfIDPvg6uqKI2xXr2"
 	crossorigin="anonymous">
+	
+	<script src="//code.jquery.com/jquery-1.11.1.min.js"></script>
+	
 <script
 	src="https://cdn.jsdelivr.net/npm/bootstrap@4.5.3/dist/js/bootstrap.bundle.min.js"
 	integrity="sha384-ho+j7jyWK8fNQe+A12Hb8AhRq26LrZ/JpcUGGOn+Y7RsweNrtN/tE3MoK7ZeZDyx"
@@ -45,7 +48,14 @@
 
 	<nav class="navbar navbar-expand-large navbar-light"
 		style="background-color: #e3f2fd;"> <a class="navbar-brand"
-		href="#">BANCO JAVA ADMIN</a>
+			<%Usuario usuarioAdmin = new Usuario();
+		
+		
+		
+		usuarioAdmin =(Usuario)session.getAttribute("UsuarioAdminLogin"); %>
+		
+		
+		href="#">BANCO JAVA ADMINISTRADOR: <%= usuarioAdmin.getNombre()%></a> 
 	<button class="navbar-toggler" type="button" data-toggle="collapse"
 		data-target="#navbarNavDropdown" aria-controls="navbarNavDropdown"
 		aria-expanded="false" aria-label="Toggle navigation">
@@ -54,6 +64,16 @@
 	<div class="collapse navbar-collapse" id="navbarNavDropdown">
 		<ul class="navbar-nav">
 			<div class="row">
+			
+			<div class="col-md-2">
+						<li class="nav-item active"><a class="nav-link" href="IndexAdmin.jsp">Home
+								<span class="sr-only">(current)</span>
+						</a></li>
+					</div>
+			
+			
+			
+			
 				<div class="col-md-2">
 					<li class="nav-item dropdown"><a
 						class="nav-link dropdown-toggle" href="#"
@@ -61,38 +81,46 @@
 						aria-haspopup="true" aria-expanded="false"> Clientes </a>
 						<div class="dropdown-menu"
 							aria-labelledby="navbarDropdownMenuLink">
-							<a class="dropdown-item" href="#">Nuevo cliente</a> <a
-								class="dropdown-item" href="#">Modificar cliente</a>
-							<div class="dropdown-divider"></div>
-							<a class="dropdown-item" href="#">Eliminar cliente</a>
+							<a class="dropdown-item" href="UsuarioDatosPersonales.jsp">Nuevo cliente</a>
+							 <a	class="dropdown-item" href="ServletUsuario?ListaClientes=1&pagina=1">Listar clientes</a>
+							
 						</div></li>
 				</div>
-				<div class="col-md-2">
-					<li class="nav-item dropdown"><a
-						class="nav-link dropdown-toggle" href="#"
-						id="navbarDropdownMenuLink" role="button" data-toggle="dropdown"
-						aria-haspopup="true" aria-expanded="false"> Cuentas </a>
-						<div class="dropdown-menu"
-							aria-labelledby="navbarDropdownMenuLink">
-							<a class="dropdown-item" href="#">Nueva cuenta</a> <a
-								class="dropdown-item" href="#">Modificar cuenta</a>
-							<div class="dropdown-divider"></div>
-							<a class="dropdown-item" href="#">Eliminar cuenta</a>
-						</div></li>
+			<div class="col-md-2">
+					
+							<a class="dropdown-item" href="ServletCuenta?ListaCuentas=1&pagina=1">Cuentas</a>
+						
 				</div>
 				<div class="col-md-3">
-					<li class="nav-item"><a class="nav-link" href="#">Solicitudes
+					<li class="nav-item"><a class="nav-link" href="PedirPrestamos.jsp">Solicitudes
 							de prestamos</a></li>
 				</div>
 				<div class="col-md-2">
 					<li class="nav-item"><a class="nav-link" href="#">Informes</a>
 					</li>
 				</div>
+				
+				<div class="col-md-2">
+						<li class="nav-item active"><a class="nav-link" href="ServletUsuario?Logout=1">Salir
+								<span class="sr-only">(current)</span>
+						</a></li>
+					</div>
+			
+				
 			</div>
 		</ul>
+		
+		
 	</div>
 	</nav>
 	<div class="container">
+	
+	<% 
+			
+			if(usuarioAdmin!=null)
+			{%>
+	
+	
 		<div class="row mt-2">
 			<div class="col-md-12">
 				<h3 class="display-4 text-info">Nomina de Clientes</h3>
@@ -105,7 +133,7 @@
 		</div>
 		<div class="row mt-2 d-flex justify-content-center"
 			style="position: relative; top: 50px">
-			<div class="col-md-6" style="">
+			<div class="col-md-10" style="">
 				<form action="ServletUsuario" method="get">
 					<label for="">Buscar</label> 
 					<input type="text"
@@ -140,7 +168,15 @@
 							<td><%=user.getPersona().getNombre()%></td>
 							<td><%=user.getPersona().getApellido()%></td>
 							<td><%=user.getPersona().getDni()%></td>
-						    <td><a class="btn btn-primary" href="ServletCuenta?nuevaCuenta=<%=user.getId()%>">Nueva Cuenta</a></td>
+						    
+						   <%if(user.getCuentas() < 3){ %>
+						    <td><a class="btn btn-primary" href="ServletCuenta?nuevaCuenta=<%=user.getId()%>" >Nueva Cuenta</a></td>
+
+							<%} else { %>
+							 <td><a class="btn btn-danger" href="ServletCuenta?nuevaCuenta=<%=user.getId()%>" disabled>Ya posee 3 cuentas</a></td>
+							<%} %>
+							
+							
 							<td><a class="btn btn-info" href="ServletUsuario?cargaModificar=<%=user.getId()%>">Modificar</a></td>
 							<td><a class="btn btn-warning" href="ServletUsuario?cargaEliminar=<%=user.getId()%>">Eliminar</a></td>
 
@@ -171,6 +207,16 @@
 			</div>
 
 		</div>
+		<%}%>
+			<%
+		          
+			
+			session.setAttribute("UsuarioAdminLogin",usuarioAdmin);	
+			
+		
+			
+			%>
+		
 	</div>
 </body>
 </html>
